Move dupe checker to its own function
