+++ packages/server/src/synchronizer/index.ts
@@ -13,7 +13,7 @@ import {remove, pathExists, ensureDir} from 'fs-extra'
 import {ciLog} from '../ciLog'
 import {runRule} from './rules-runner'
 import through from 'through2'
-import chalk from 'chalk'
+import {checkDuplicateRoutes} from './check-duplicate-routes'
 
 type SynchronizeFilesInput = {
   src: string
@@ -54,32 +54,13 @@ export async function synchronizeFiles({
   }
 
   const entries = fg.sync(includePaths, {ignore: options.ignored, cwd: options.cwd})
+
+  checkDuplicateRoutes(entries)
+
   const manifest = Manifest.create()
   const {stream, watcher} = watch(includePaths, options)
 
   await clean(destPath)
-
-  const allRoutes = entries.filter((route) => route.includes('routes'))
-  const cleanRoutes = allRoutes.map((route) => route.split('routes/')[1])
-  const duplicateRoutes = cleanRoutes.filter((e, i, a) => a.indexOf(e) !== i)
-  const duplicateRoutesIndexes = duplicateRoutes.map((route) =>
-    cleanRoutes.flatMap((cleanRoute, i) => (cleanRoute === route ? i : [])),
-  )
-  const duplicateRoutesPaths = duplicateRoutesIndexes.map((indexes) =>
-    indexes.map((index) => allRoutes[index]),
-  )
-
-  if (duplicateRoutesPaths.length > 0) {
-    console.log(chalk.yellow('Warning: You have created conflicting routes: \n'))
-
-    duplicateRoutesPaths.forEach((route) => {
-      route.forEach((path) =>
-        console.log(`- ${path.split('routes/')[0]}routes/${chalk.bold.redBright(path.split('routes/')[1])}`),
-      )
-    })
-
-    process.exit(0)
-  }
 
   return await new Promise((resolve, reject) => {
     pipeline(
